requirement.txt
streamlit
langchain
openai
faiss-cpu
pandas
python-dotenv
pymupdf
